Data Flow

+++ mac15wr/mvc/model/database/WRRealmCourse.h
@@ -9,12 +9,12 @@
 #import <Realm/Realm.h>
 
 @interface WRRealmCourse : RLMObject
-@property double course_id;
+@property NSInteger course_id;
 @property NSString* sis_course_id;
 @property NSString* title;
-@property double min_units;
-@property double max_units;
-@property double total_max_units;
+@property NSInteger min_units;
+@property NSInteger max_units;
+@property NSInteger total_max_units;
 @property NSString* desc;
 @property NSString* diversity_flag;
 @property NSString* effective_term_code;
